ready exept skeletop and active
