+++ src/blocks/Main.js
@@ -2,13 +2,14 @@ import React, { useState } from 'react'
 import SideBar from './SideBar'
 import Nav from './Nav'
 import CenterBlock from './CenterBlock'
+import { PlayLists } from './Const'
 
 function Main() {
   return (
     <main className="main">
       <Nav />
       <CenterBlock />
-      <SideBar />
+      <SideBar {...PlayLists} />
     </main>
   )
 }
