+++ src/css/style.css
@@ -922,3 +922,9 @@ body {
   border-radius: 100px;
   border: 8px solid #4e4e4e;
 }
+
+.active {
+  border-color: #ad61ff;
+  color: #ad61ff;
+  cursor: pointer;
+}
